requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black 
celery==5.2.7
certifi
click
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
convertdate==2.4.0
Django==4.1.6
django-environ==0.9.0
flit_core
kombu==5.2.4
lunardate==0.2.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
paho-mqtt==1.6.1
pathspec
platformdirs
prompt-toolkit==3.0.36
pyluach==2.0.2
PyMeeus==0.5.11
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
tomli
typing==3.7.4.3
typing_extensions
vine==5.0.0
wcwidth==0.2.6
workalendar==16.4.0
xlwt==1.3.0
